requirements.txt
openai
python-dotenv
langgraph
pyairtable
langchain-core
langchain-openai
python-docx
pypdf
fastapi
uvicorn
requests
langchain-google-genai
langchain-huggingface
langchain-text-splitters
supabase
neo4j
langchain
langchain-community
langchain-google-genai
langchain-huggingface
sentence-transformers
pydantic
pytest
graphiti-core[google-genai]

